Move internal packages to pkg directory and fix linter issues
